requirements.txt
streamlit
mysql-connector-python==9.1.0
mysqlclient==2.2.6
SQLAlchemy==2.0.36